move column and segment selection to FilterColumnPicker
