+++ frontend/src/metabase/common/components/FilterPicker/FilterColumnPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterColumnPicker";
